Add post-stow hooks for vim and zsh rules
These make sure to install or update any plugins after stowing
everything. For now, it installs/updates fzf, zplug and vim-plug. This
required one small change to my .vimrc: bringing up the previous
workspace through the vim-ctrlspace plugin was making vim exit with an
error code, so this had to be disabled.