BREAK CHANGE: Refactor vim configuration
Moved everything out from `.vimrc` into a plugin structure within the
autoload directory. Most of the plugin configuration lives in
`.vim/rc/dein.toml`. Key mappings are meant to be configured in a JSON
file similar to that of Atom's. Furthermore, being a big fan of
Spacemac's key mappings and their groupings, the configuration is
intended to mimic that.