feat(neovim): upgrade to new which-key spec
We're now conforming to the v3 spec for plugins that which-key expects.
One deviation that is present in the way we're loading the which-key
plugin that differs from others is that we have to load its main module
in a protected call. Doing so prevents the entire startup routine from
failing to load the rest of Makyo and actually loads the module,
something that Aniseed's `module` or a regular call to `require` fail to
do.