fix(neovim): Make <Enter> work in insert mode
There appeared to be a conflict when pressing <Enter> in insert mode
caused by the coc-snippets COC extension. The fix involved calling
`coc#pum#confirm()` instead as the old function being referenced looks
to be deprecated. Additionally, we explicitly set the shortcuts for
jumping around snippets here as they were not working.